+++ application/modules/shop/controllers/BackendProductController.php
@@ -32,7 +32,7 @@ use yii\web\NotFoundHttpException;
 use yii\web\ServerErrorHttpException;
 use app\backend\actions\MassPublishAction;
 use app\backend\actions\CategoryMovementsAction;
-use app\modules\shop\actions\PacketPriceEditAction;
+use app\modules\shop\actions\BatchEditPriceAction;
 
 class BackendProductController extends BackendController
 {
@@ -139,8 +139,8 @@ class BackendProductController extends BackendController
             'categoryMovements' => [
                 'class' => CategoryMovementsAction::className(),
             ],
-            'packet-price-edit' => [
-                'class' => PacketPriceEditAction::className(),
+            'batch-edit-price' => [
+                'class' => BatchEditPriceAction::className(),
             ]
         ];
     }
